reduce USE_AVB macro definitions
The purpose of this patch is try to reduce the USE_AVB
macro definitions in kernelflinger to make it more easy
for maintenance.
The major change is to split the android.c and security.c
into multiple files, move the verified boot related
functions into android_vb1.c and android_vb2.c, vb1 suffix
means verified boot 1.0 specific code, vb2 suffix means
verified boot 2.0 specific code, same for security.c.

Tracked-On: https://jira.devtools.intel.com/browse/OAM-75485
Reviewed-on: https://android.intel.com:443/664254
