+++ include/libkernelflinger/security.h
@@ -34,6 +34,8 @@
 #include <efi.h>
 #include <efilib.h>
 #include <openssl/x509.h>
+#include "android_vb.h"
+#include "security_vb.h"
 
 #ifndef _SECURITY_H_
 #define _SECURITY_H_
@@ -44,37 +46,6 @@
 
 #define SETUP_MODE_VAR	        L"SetupMode"
 #define SECURE_BOOT_VAR         L"SecureBoot"
-
-/* Compute sums of the public key value of X509 input CERT */
-EFI_STATUS pub_key_sha256(X509 *cert, UINT8 **hash_p);
-EFI_STATUS pub_key_sha1(X509 *cert, UINT8 **hash_p);
-
-/* Given an Android boot image, test if it is signed with the provided
- * certificate or the embedded one
- *
- * Parameters:
- * bootimage - data pointer to an Android boot image which may or may not
- *             be signed. This code may seek up to BOOT_SIGNATURE_MAX_SIZE
- *             past the end of the boot image size as reported by its header
- *             to search for the ASN.1 AndroidVerifiedBootSignature message.
- * der_cert  - DER certificate to validate image with
- * cert_size - Size of DER certificate
- * target    - Pointer to buffer of BOOT_TARGET_SIZE, which will be filled in
- *             with AuthenticatedAttributes 'target' field iff the image is
- *             verified. Caller should only check this on EFI_SUCCESS.
- * verifier_cert  - Return the certificate that validated the boot image
- *
- * Return values:
- * BOOT_STATE_GREEN  - Boot image is validated against provided certificate
- * BOOT_STATE_YELLOW - Boot image is validated against embedded certificate
- * BOOT_STATE_RED    - Boot image is not validated
- */
-UINT8 verify_android_boot_image(
-        IN VOID *bootimage,
-        IN VOID *der_cert,
-        IN UINTN cert_size,
-        OUT CHAR16 *target,
-        OUT X509 **verifier_cert);
 
 BOOLEAN is_platform_secure_boot_enabled(VOID);
 BOOLEAN is_eom_and_secureboot_enabled(VOID);
@@ -92,12 +63,6 @@ EFI_STATUS verify_pkcs7(const unsigned char *cert_sha256, UINTN cert_size,
 			VOID **data, int *size);
 #endif  /* BOOTLOADER_POLICY */
 
-/* Given a X509 certificate, build the following string:
- * COMMON_NAME:#PUBLIC_KEY_SHA1
- * Where COMMON_NAME is the certificate issuer CN and PUBLIC_KEY_SHA1
- * is the X509 certificate public key SHA1 hash.
- */
-EFI_STATUS get_android_verity_key_id(X509 *cert, char **value);
 
 /* Structure for RoT info (fields defined by Google Keymaster2)
 */
@@ -128,14 +93,15 @@ struct rot_data_t{
 } ;
 
 /* Initialize the struct rot_data for startup_information */
-#ifdef USE_AVB
-EFI_STATUS get_rot_data(IN VOID * bootimage, IN UINT8 boot_state,
-                        IN const UINT8 *pub_key,
-                        IN UINTN pub_key_len,
-                        OUT struct rot_data_t *rot_data);
+EFI_STATUS get_rot_data(
+        IN VOID *bootimage,
+        IN UINT8 boot_state,
+        IN VBDATA *vb_data,
+        OUT struct rot_data_t *rot_data);
 
-#else
-EFI_STATUS get_rot_data(IN VOID *bootimage, IN UINT8 boot_state, IN X509 *verifier_cert,
-                        OUT struct rot_data_t *rot_data);
-#endif
+EFI_STATUS raw_pub_key_sha256(
+        IN const UINT8 *pub_key,
+        IN UINTN pub_key_len,
+        OUT UINT8 **hash_p);
+
 #endif
