+++ libqltipc/ql-tipc/Android.mk
@@ -19,6 +19,11 @@ LOCAL_EXPORT_C_INCLUDE_DIRS := $(LOCAL_PATH)/../../include/libqltipc
 LOCAL_C_INCLUDES := $(LOCAL_PATH)/include \
                     $(LOCAL_PATH)/../interface/include \
                     $(LOCAL_PATH)/../../include
+
+ifeq ($(BOARD_AVB_ENABLE),true)
+LOCAL_C_INCLUDES += $(LOCAL_PATH)/../../avb
+endif
+
 LOCAL_SRC_FILES := \
 	ipc.c \
 	ipc_dev.c \
